Inform event handlers on why they are being polled
Most use cases of the event loop are based around nonblocking sockets.
Up until now, it has been fine to just read a socket, even if we know
it's empty

In some cases we (maybe) are not so lucky. In our window manager we open
a drm file handle and we have complete control of when it will fire
events for us. Maybe it could have been set to non-blocking, but I
didn't feel like looking into it

In this case reading the file descriptor in our initial case is wrong.
In most cases it is still correct as epoll will only trigger on edge, so
we need to consume all bytes from a socket before waiting.

Inform pollers of why they are polling. This allows some consumers of
events to opt out of reading unless we are confident that there is data
to read.