Dynamic GL loading
Up until now, we've assumed that by including gl.h with
GL_GLEXT_PROTOTYPES defined we didn't have to use a fancy OpenGL loader.
It turns out that this is only true because the GL lib on my systems
are provided by libglvnd, which happens to define these functions
unconditionally.

On some systems (e.g. an alpine VM i'm testing on with mesa installed)
this is not the case. Linking fails on these systems as the extension
functions are only available through dynamic loading.

Remove the GL_GLEXT_PROTOTYPE macros and add a dynamic loader.

The motivating use case for this is in a window manager we're working on
that uses GL_OES_EGL_image to import textures from other processes. This
means that importing only core APIs is not sufficient, but only in this
application. Allow library consumers to specify extra OpenGL extensions
for bindings generation

High level details:
* Yoink xml definition of opengl functions from
  https://github.com/KhronosGroup/OpenGL-Registry/blob/main/xml/gl.xml
* Add a sphxml library that is inspired by an xml parser stolen from
  sphaerophoria/sphwayland-client@aea0314cd05bb02579bb1ea9a85f7f10602d77a3
  but with a couple fixes and refactor to work with std.Io.Reader
* Parse gl.xml to find all functions associated with gl core apis and
  requested extensions
* Append to translate-c of gl.h with global variables corresponding to
  each imported function
* Add a load function to initialize the globals

Implementation notes:
* gl_extensions needs to make it through a lot of build scripts. If we
  use a different set of extensions, we end up with a different
  sphrender module which will conflict. Everyone who imports sphrender
  needs to import with the exact same set of extensions
* Add StringHashMapLinear for use in inject_loader. New build time exe
  manages to get away with a single BufAllocator. Using our hash map
  instead of std is desirable here, but it didn't expose string key
  support yet