Add RuntimeSegmentedListUnmanaged
This type gets stored often next to other RSLs, storing the same
allocator multiple times is silly. Add an unmanaged variant

In order to preserve the API, both the unmanaged and managed versions
store their own data. They call out to RSLImpls to do the work. This is
an attempt to keep the diff small. RSLImpl and RSLImplConst should be
very similar to the previous code