Add way to upgrade LinearAllocator into LinearAllocators
A frequent pattern we use is to create a single BufAllocator, use one
end for long lived allocations, and the other for scratch space. This
works well for any application where all lifetimes have defined scopes.

A problem arises though when we are in a scenario where we have a
scratch allocator, and we want to call a function that does some scratch
work before returning. E.g.

```
fn someFn(ret_alloc: std.mem.Allocator, scratch: sphtud.alloc.LinearAllocator) ![]const u8
```
In this scenario we cannot forward our scratch AND use it as scratch
space. Sometimes we can dodge the problem if we do all return
allocations up front, then it's ok if ret_alloc and scratch are actually
the same allocator, but sometimes this is not desirable. E.g. any
incremental parsing that also involves temporary allocations. We do not
know the length of the return until we get to the end of the function.

Add a way to upgrade a LinearAllocator into a double ended one. This
will allocate the maximum space available, and create 2 linear
allocators from that newly allocated space. All allocations from the
original BufAllocator will now fail until we restore a checkpoint or
commit the allocators we just made.

Add a way to commit the allocations from the above allocators back into
where they came from. The pattern for calling someFn is now...

```
const val = blk: {
    const tmp = try scratch.makeDoubleEnded();
    defer scratch.commitDoubleEnded(tmp);
    break :blk try someFn(tmp.preserve.allocator(), tmp.discard);
};
```