Add way to unpage memory used by BufAllocator
If we create a buf allocator around a large buffer, we are relying on
virtual memory to not page in that buffer until it is touched. The
problem is that if on one iteration of our main loop we do a large
allocation, that memory is forced to be tracked by the kernel until our
app closes.

We know that we do not need this memory. Let the kernel know too

To avoid unnecessary page fault spam, only release memory if it hasn't
been used for a few iterations. For most use cases our usage should be
fairly static.

Expected usage:
* Top level BufAllocator passes in a tracer
* On each iteration of the main loop tracer asks for memory reclamation
* If memory has not been used for a few iterations it gets unpaged
* On future allocations if we grow, we'll page fault and those pages
  will get brought back in