Use runtime expansion allocator params in RuntimeSegmentedList
Originally the RuntimeSegmentedList was written specifically with our
tiny page allocator in mind. This meant that all expansions had to be at
least 256 bytes large. Eventually it became clear that this is a very
useful type for scratch space and linear allocators as well. This use
has slightly different constraints

To solve this we added some comptime parameters to adjust the minimum
page size of the expansion allocator, and to specify if we should
release allocated blocks of memory when we're done with them

This ends up being quite annoying when this type gets nested in other
types. The comptime parameters infect the whole stack. It turns out that
there isn't a very noticeable difference if we just make this a runtime
parameter

Implementation details:
* Add concept of ExpansionAlloc in sphtud.util. This exposes all the
  same information as before but with runtime types
* Update RSL to take ExpansionAlloc instead of a normal alloc for all
  functions that are allocating after init
* Rename RSL (and all wrappers) to remove the
  Configurable/Linear/Sphalloc distinction. These parameters now come
  from the ExpansionAlloc
* Add expansion() functions to our relevant allocators for ease of use
* Move sphevent.Loop and friends back to concrete type
* initSingleAlloc() removed as we need the expansion alloc to size our
  initial list alloc
* Unmanaged needs to hold the min expansion size now, there might be
  some clever way to avoid the extra u8, but this is probably fine for
  now
  * We could take in the expansion info on every call, even ones that
    don't allocate, but that seems more confusing

Bonus changes:
* TinyPageAllocator was holding an arbitrary allocator when it always
  wants to use the system page allocator
* HashMap was holding an arena that it never used. This became clear
  after the type change